refactor(extensions)!: move Mermaid extension from yfm to additional folder
