+++ demo/Playground.tsx
@@ -20,9 +20,9 @@ import {
 } from '../src';
 import type {EscapeConfig, ToolbarActionData} from '../src/bundle/Editor';
 import {Extension} from '../src/cm/state';
+import {Mermaid} from '../src/extensions/additional/Mermaid';
 import {FoldingHeading} from '../src/extensions/yfm/FoldingHeading';
 import {Math} from '../src/extensions/yfm/Math';
-import {Mermaid} from '../src/extensions/yfm/Mermaid';
 import {YfmHtmlBlock} from '../src/extensions/yfm/YfmHtmlBlock';
 import {getSanitizeYfmHtmlBlock} from '../src/extensions/yfm/YfmHtmlBlock/utils';
 import {cloneDeep} from '../src/lodash';
