+++ src/extensions/additional/Mermaid/MermaidNodeView/Mermaid.scss
