+++ src/extensions/additional/Mermaid/MermaidNodeView/NodeView.tsx
