+++ src/extensions/additional/Mermaid/MermaidNodeView/index.ts
