+++ src/extensions/additional/Mermaid/MermaidSpecs/const.ts
