+++ src/extensions/additional/Mermaid/const.ts
